# backend/api.yaml
...
openapi: 3.0.3
info:
  title: Katalyst Placement Portal API
  description: |
    Decoupled Flask JSON API for Katalyst V2.

    - Auth: `Authorization: Bearer <access_token>` (JWT)
    - Roles: `admin`, `student`, `company` (blacklisted users get 403)
    - UI wording uses "drives"; API resources use `job_positions`
    - Errors: `{ "error", "message", "status" }`
  version: 2.0.0
  contact:
    name: Deva Vasista
    email: 21f2000226@ds.study.iitm.ac.in
servers:
  - url: http://127.0.0.1:3000
    description: Local Flask API (`cd backend && python app.py`)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from POST /api/auth/login
    RefreshAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Refresh token from login (use only on /api/auth/refresh)
  schemas:
    Error:
      type: object
      required: [error, message, status]
      properties:
        error: { type: string, example: Bad Request }
        message: { type: string }
        status: { type: integer, example: 400 }
...
    TaskQueued:
      type: object
      properties:
        message: { type: string }
        task_id: { type: string }
        delivery: { type: string, enum: [download, email, both] }
    TaskStatus:
      type: object
      properties:
        task_id: { type: string }
        state:
          type: string
          enum: [PENDING, STARTED, SUCCESS, FAILURE, RETRY]
        result: { type: object, additionalProperties: true }
        error: { type: string }
  responses:
...
    Forbidden:
      description: Wrong role or blacklisted
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /api/jobs/tasks/{task_id}:
    get:
      tags: [Jobs]
      summary: Poll Celery task state
      security: [{ BearerAuth: [] }]
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Task status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskStatus" }
  /api/jobs/exports/applications:
    post:
      tags: [Jobs]
      summary: Queue applications CSV export
      description: Student or company only. Email delivery uses SMTP (MailHog locally).
      security: [{ BearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                delivery:
                  type: string
                  enum: [download, email, both]
                  default: download
      responses:
        "202":
          description: Queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskQueued" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/jobs/exports/{filename}:
    get:
      tags: [Jobs]
      summary: Download finished export CSV
      security: [{ BearerAuth: [] }]
      parameters:
        - name: filename
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema: { type: string, format: binary }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/jobs/admin/daily-reminders:
    post:
      tags: [Jobs]
      summary: Queue daily deadline and interview reminder emails
      security: [{ BearerAuth: [] }]
      responses:
        "202":
          description: Queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskQueued" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/jobs/admin/monthly-report:
    post:
      tags: [Jobs]
      summary: Queue monthly HTML activity report email
      security: [{ BearerAuth: [] }]
      responses:
        "202":
          description: Queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskQueued" }
        "403": { $ref: "#/components/responses/Forbidden" }